requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
boto3
python-dotenv
requests
astronomer-cosmos
dbt-snowflake
sec-api
pandas
snowflake-snowpark-python
boring-semantic-layer[mcp]
ibis-framework[snowflake]